feat[array]: refactor `zip` compute kernel to lazy `zip` via ScalarFnArray
