+++ encodings/runend/src/compute/fill_null.rs
@@ -4,7 +4,7 @@
 use vortex_array::ArrayRef;
 use vortex_array::IntoArray;
 use vortex_array::builtins::ArrayBuiltins;
-use vortex_array::compute::FillNullReduce;
+use vortex_array::expr::FillNullReduce;
 use vortex_error::VortexResult;
 use vortex_scalar::Scalar;
