+++ encodings/sparse/src/rules.rs
@@ -4,8 +4,8 @@
 use vortex_array::ArrayRef;
 use vortex_array::IntoArray;
 use vortex_array::builtins::ArrayBuiltins;
-use vortex_array::compute::NotReduce;
-use vortex_array::compute::NotReduceAdaptor;
+use vortex_array::expr::NotReduce;
+use vortex_array::expr::NotReduceAdaptor;
 use vortex_array::optimizer::rules::ParentRuleSet;
 use vortex_error::VortexResult;
