+++ vortex-array/src/arrays/bool/vtable/kernel.rs
@@ -4,7 +4,7 @@
 use crate::arrays::BoolVTable;
 use crate::arrays::TakeExecuteAdaptor;
 use crate::arrays::filter::FilterExecuteAdaptor;
-use crate::compute::FillNullExecuteAdaptor;
+use crate::expr::FillNullExecuteAdaptor;
 use crate::kernel::ParentKernelSet;
 
 pub(super) const PARENT_KERNELS: ParentKernelSet<BoolVTable> = ParentKernelSet::new(&[
