+++ vortex-array/src/arrays/chunked/compute/rules.rs
@@ -12,8 +12,8 @@ use crate::arrays::ChunkedVTable;
 use crate::arrays::ConstantArray;
 use crate::arrays::ConstantVTable;
 use crate::arrays::ScalarFnArray;
-use crate::compute::FillNullReduceAdaptor;
-use crate::compute::ZipReduceAdaptor;
+use crate::expr::FillNullReduceAdaptor;
+use crate::expr::ZipReduceAdaptor;
 use crate::optimizer::ArrayOptimizer;
 use crate::optimizer::rules::ArrayParentReduceRule;
 use crate::optimizer::rules::ParentRuleSet;
