+++ vortex-array/src/arrays/decimal/compute/fill_null.rs
@@ -19,7 +19,7 @@ use crate::IntoArray;
 use crate::ToCanonical;
 use crate::arrays::DecimalVTable;
 use crate::arrays::decimal::DecimalArray;
-use crate::compute::FillNullKernel;
+use crate::expr::FillNullKernel;
 use crate::validity::Validity;
 use crate::vtable::ValidityHelper;
